data_types.OpticsTypes -> rte.OpticsTypes
